+++ braintree/transaction.py
@@ -20,6 +20,7 @@ from braintree.configuration import Configuration
 from braintree.credit_card import CreditCard
 from braintree.customer import Customer
 from braintree.paypal_account import PayPalAccount
+from braintree.paypal_here import PayPalHere
 from braintree.europe_bank_account import EuropeBankAccount
 from braintree.subscription_details import SubscriptionDetails
 from braintree.resource_collection import ResourceCollection
@@ -700,6 +701,8 @@ class Transaction(Resource):
             self.credit_card_details = CreditCard(gateway, attributes.pop("credit_card"))
         if "paypal" in attributes:
             self.paypal_details = PayPalAccount(gateway, attributes.pop("paypal"))
+        if "paypal_here" in attributes:
+            self.paypal_here_details = PayPalHere(gateway, attributes.pop("paypal_here"))
         if "local_payment" in attributes:
             self.local_payment_details = LocalPayment(gateway, attributes.pop("local_payment"))
         if "europe_bank_account" in attributes: